# Crypt-Creeper_Api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crypt Creeper",
        "description": "API functions for Crypt Creeper.",
        "contact": {
            "email": "glugsludge@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/create": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Creates a new user.",
                "description": "Create the specified resource.\nCreates a new user.",
                "operationId": "58a1e7e5ea2085144e7709685a657ed4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "Name for the new user",
                                "value": "abbytorade"
                            }
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "The email for the new user",
                                "value": "abbytorade@icloud.com"
                            }
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password for the new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "The password for the user",
                                "value": "Ud345678"
                            }
                        }
                    },
                    {
                        "name": "faction_id",
                        "in": "path",
                        "description": "ID of the faction to join.",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "int": {
                                "summary": "The id for the faction to join",
                                "value": "6"
                            }
                        }
                    },
                    {
                        "name": "profile_pic",
                        "in": "path",
                        "description": "Profile picture for the new user",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "string": {
                                "summary": "The picture in base64",
                                "value": "(base64 picture)"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User was created."
                    },
                    "422": {
                        "description": "Info entered was invalid."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Creates a new token for the user.",
                "description": "Create the specified resource.\nCreates a new token for the user.",
                "operationId": "9f6f4aca790a5229fb638e798623c6e0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "Name of the user",
                                "value": "abbytorade"
                            }
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "The password for the user",
                                "value": "Ud345678"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Token was created."
                    },
                    "422": {
                        "description": "Info entered was invalid."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/getUserData": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Retrieves all the data from the logged in user.",
                "description": "Retrieves the specified resource.\nGets all the data from the logged in user.",
                "operationId": "8eb5171f290d295a105473531b117d9d",
                "responses": {
                    "200": {
                        "description": "All data from the user"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/getTop8": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Retrieves the top 8 plays from the user.",
                "description": "Retrieves the specified resource.\nGets the best 8 scores from the user.",
                "operationId": "0b21c163e41ba039b2076ba1b0351c5d",
                "responses": {
                    "200": {
                        "description": "List of plays from the user"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Logs out the user by deleting the token used.",
                "description": "Destroys the specified resource.\nLogs out an user by deleting the token used.",
                "operationId": "d52b7004aad9cc11ac927b616ad9c1f2",
                "responses": {
                    "200": {
                        "description": "The token was deleted"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/change-name": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Changes the name of the user.",
                "description": "Changes the specified resource.\nChanges the username of the logged in user.",
                "operationId": "81503a646abdc1d4c6420f56856c9dd9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "New name for the user",
                                "value": "abbytorade"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Name was changed successfully "
                    },
                    "400": {
                        "description": "Name is not valid"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
        "/api/user/change-password": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Changes the name of the user.",
                "description": "Changes the specified resource.\nChanges the password of the logged in user.",
                "operationId": "157894b6e5270c75e20d2661b242e883",
                "parameters": [
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "Old password",
                                "value": "Ud345678"
                            }
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "path",
                        "description": "New pasword for the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "New password",
                                "value": "876543dU"
                            }
                        }
                    },
                    {
                        "name": "repeat_password",
                        "in": "path",
                        "description": "Repeat new pasword for the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "New password (repeated)",
                                "value": "876543dU"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Name was changed successfully "
                    },
                    "400": {
                        "description": "Password is not valid"
                    },
                    "401": {
                        "description": "Passwords do not match"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
...
    }
}